feat(ydr): render gizmos for capsule lights
Light rendering in Blender is still the same as point lights, but the
gizmo shows the capsule bounds which should be closer to how it
appears in-game.

feat(ydr): interactive capsule light gizmos

Capsule lights can now be resized in the 3D view via a cage gizmo
that appears when hovering the active capsule light, similar to
how builtin light gizmos work.

fix(ydr): incorrect scaling of capsule light gizmos

Was using the radius in places where the diameter was actually
needed, causing the capsule gizmo to be smaller than it should be.